% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineresult.R
\name{combineresult}
\alias{combineresult}
\title{combining and formatting the individual PA results}
\usage{
combineresult()
}
\value{
TRUE when successful, FALSE otherwise
}
\description{
It yields _STATS.RData, summarizing and formatting the intermediate outputs such as true positive rates (TPR), true negative rates (TNR), false positive rates (FPR) and false negative rates (FNR), plus the Youden's best p-value threshold, sensitivity (TPR), specificity (TNR), accuracy, precision and recall.
}
\details{
This function takes "*_SUM.RData" generated by pathwayko, combines them
	into a single object which can be further analyzed or plotted. All "_SUM"
		data should be analyzed by the same sets of PA methods, if not by identical
		PA methods under identical parameters. (i.e. with same KEGG xmls, same criteria
	for choosing DE genes, etc)
}
