% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceplot.R
\name{evidenceplot}
\alias{evidenceplot}
\title{generating two-dimensional evidence plot}
\usage{
evidenceplot()
}
\description{
It generates two-dimensional evidence plot for methods based on or derived from SPIA
}
\details{
Specifically, this function looks for two columns in raw pathway analysis results:
	"pPERT" and "pNDE", to plot the evidence plot. 
		Reference: Tarca A. L. et al. (2009) A novel signaling pathway impact analysis. Bioinformatics, 25, 75–82.)
}
