% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtertrue.R
\name{filtertrue}
\alias{filtertrue}
\title{extracting the list of true positive KO KEGG pathways}
\usage{
filtertrue()
}
\description{
It extracts the list of true positive KO KEGG pathways that contain and are impacted by the KO gene from the results of each method
}
\details{
This function scans user provided directory for any .csv files under the current directory
	recursively and reads in all .csv files not tagged with control keywords like "AUC" or "pathway".
	Hence this function should only be used on directories with no .csv files from other sources.
	After data input, this function filters out from all PA results entries of true positive KEGG pathways
		in that analysis and saves each of filtered pathway analysis result in a separate .csv file.
}
