% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highedges.R
\name{highedges}
\alias{highedges}
\title{calculating edge scores, determining the change-point and generating the KO-gene associated sub-graph}
\usage{
highedges(
  KEGGgraph = NULL,
  DataObject = NULL,
  condition_FC = NULL,
  condition_pVal = NULL,
  mc.cores = 1
)
}
\arguments{
\item{KEGGgraph}{graphical representation of pathways in KEGG}

\item{DataObject}{Object generated by calling preprocess.R}

\item{condition_FC}{named list of logFC of all genes in DataObject}

\item{condition_pVal}{named list of adjusted pval of all genes in DataObject}

\item{mc.cores}{number of cores to be used in parallel lapply}
}
\value{
various information grouped together as a list
}
\description{
It is not executed itself, instead, it is invoked by pathwayko.
}
\details{
Reference: Hanoudi S., et al. (2017) Identifying biologically relevant putative mechanisms in a given phenotype comparison. PLoS ONE 12, e0176950.
}
\author{
Original implementation by Samer Hanoudi and Sorin Draghici, modification by Hannan Ai
}
