% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSPIAdata.R
\name{makeSPIAdata}
\alias{makeSPIAdata}
\title{parsing KEGG xmls ready for pathway enrichment analysis}
\usage{
makeSPIAdata(kgml.path=NULL,organism="mmu",out.path=".")
}
\arguments{
\item{kgml.path}{path to KEGG xml files}

\item{organism}{three letter organism code from KEGG database}

\item{out.path}{path to which generated RData file will be placed}
}
\value{
TRUE on success, FALSE otherwise
}
\description{
Reference: Tarca A. L. et al. (2009) A novel signaling pathway impact analysis. Bioinformatics, 25, 75–82.
}
\author{
Original implementation by Adi Laurentiu Tarca <atarca AT med.wayne.edu>, Purvesh Kathri
   <purvesh AT cs.wayne.edu> and Sorin Draghici <sorin AT wayne.edu>, source code shown at the bottom.
   Modification by Hannan Ai
}
