% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayko.R
\name{pathwayko}
\alias{pathwayko}
\title{performing the KO pathway enrichment analysis on single dataset}
\usage{
pathwayko()
}
\arguments{
\item{batch}{boolean, should only be set to TRUE by pathwayko_batch function}

\item{batch_DataObject}{should only be set by pathwayko_batch function}

\item{batch_uip}{should only be set by pathwayko_batch function}
}
\value{
NULL if encountered error, TRUE when batch is false and an internal 
	object with results if batch is true
}
\description{
It invokes utilities, highedges, functions and methods to calculate the scores of all edges (Hanoudi et al., 2017) in a known global KEGG graph (Zhang and Wiemann, 2009) to yield the distribution of edge scores, and thus automatically determines a change point on the distribution of edge scores by the change-point analysis method (Killick and Eckley, 2014). Outputs include the KO gene-associated sub-graph, the list of differential expression (DE) genes, and the list of true positive KO KEGG pathways.
		References: Hanoudi S., et al. (2017) Identifying biologically relevant putative mechanisms in a given phenotype comparison. PLoS ONE 12, e0176950; Zhang J. D. and Wiemann S. (2009) KEGGgraph: a graph approach to KEGG pathway in R and bioconductor. Bioinformatics, 25, 1470–1471; Killick R. and Eckley I. (2014) changepoint: An R package for changepoint analysis. J. Statistic Software. 58,1–19.
}
