% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayko.R
\name{pathwayko_batch}
\alias{pathwayko_batch}
\title{batch-performing the KO pathway enrichment analysis on multiple datasets}
\usage{
pathwayko_batch()
}
\value{
TRUE if all process completed successfully, FALSE otherwise
}
\description{
It conducts the KO pathway enrichment analysis on multiple datasets in a pipeline manner.
}
\details{
This function scans all directory under the working directory to identify
data object generated by preprocess function and apply pathwayko to all said objects
	using same sets of parameters obtained from the user
}
