% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayko.R
\name{pathwayko_demo}
\alias{pathwayko_demo}
\title{for pathwayko_demo on 10 KO GEO datasets}
\usage{
pathwayko_demo()
}
\value{
TRUE if all process completed successfully, FALSE otherwise
}
\description{
It generates demo results for testing on the 10 KO GEO datasets
}
