% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayview.R
\name{pathwayview}
\alias{pathwayview}
\title{rendering the true positive KO KEGG pathways by the up- and down-regulated DE genes}
\usage{
pathwayview()
}
\value{
TRUE on success, FALSE otherwise
}
\description{
It renders the true positive KO KEGG pathways by the up- and down-regulated DE genes, 
 which contain and are impacted by the KO genes. 
Reference: Luo, W. and Brouwer C., Pathview: an R/Bioconductor package for pathway-based data integration 
	and visualization. Bioinformatics, 2013, 29(14): 1830-1831, doi: 10.1093/bioinformatics/btt285
}
