% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocessing GSE data from CEL files and series matrix file}
\usage{
preprocess()
}
\value{
TRUE when successful, FALSE otherwise
}
\description{
This function reads in relevant CEL files (from *_RAW.tar)
	file and series matrix file (*_series_matrix.txt.gz) typically obtained
from NCBI database for a given GSE. The preprocessing of these data
relies on the user to interactively provide necessary information
	as such information is difficult to obtain prior to execution and hence 
to automate. Reference: Carvalho B.S. and Irizarry R.A. (2010) A framework for 
 oligonucleotide microarray preprocessing. Bioinformatics, 16, 2363–2367; 
 Ritchie M.E., et al. (2015) limma powers differential expression analyses for 
 RNA-sequencing and microarray studies. Nucleic Acids Res, 43, e47.
}
\details{
To begin, the user is expected to have obtained '*_RAW.tar' and
	'*_series_matrix.txt.gz' files of a given GSE. The user will first be
asked to choose from files under the working directory with matching 
suffix as input files. Then the user will be asked to provide relevant
information such as keywords for case/control and KO gene names to 
build experiment design for packages like 'oligo' and 'limma'. The
final resault will be saved as an compressed R object to be read in
and used by other parts of the package.
}
