% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roctest.R
\name{roctest}
\alias{roctest}
\title{conducting ROC-curve based statistics analysis}
\usage{
roctest()
}
\description{
It performs roctest on two ROC curves
}
\details{
Reference: Robin X. et al. (2011) pROC: an open-source package for R and S+ to analyze and 
 compare ROC curves. BMC Bioinformatics, 12, 77.
}
