% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplot.R
\name{violinplot}
\alias{violinplot}
\title{generating violinplots based on the combined PA results}
\usage{
violinplot()
}
\description{
It displays 12 key metrics including the Youden's best p-value threshold, specificity
 (i.e., TNR), sensitivity (i.e., TPR), FDR, FPR, FNR, accuracy, precision, recall, AUC, pAUC_SP and 
 pAUC_SE across methods under study when benchmarked on a set of KO GEO datasets.
}
