% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxtest.R
\name{wilcoxtest}
\alias{wilcoxtest}
\title{performing pairwise comparisons on the PA results}
\usage{
wilcoxtest()
}
\description{
It conducts the "paired", pairwise comparisons between each pair of two methods under study 
 when benchmarked on the same set of KO datasets. pairwise.wilcox.test is from the R `stats` package, 
 with corrections (BH FDR adjustment of p-value) for multiple testing under the two-sided mode.
}
